+++ CHANGELOG.md
@@ -9,6 +9,11 @@ Unreleased section should follow [Release Toolkit](https://github.com/newrelic/r
 
 ## Unreleased
 
+### enhancements
+- Added Query Performance Monitoring support for RDS
+- Added QueryMonitoringOnly flag to enable only query monitoring feature
+- Made individual query filtering case-insensitive
+
 ## v1.15.0 - 2025-04-21
 
 ### 🚀 Enhancements